% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcpa.R
\name{mcpa_search}
\alias{mcpa_search}
\title{Search annexation records by keyword}
\usage{
mcpa_search(query)
}
\arguments{
\item{query}{Character string to search for}
}
\value{
tibble (same schema as mcpa_annexations)
}
\description{
Search annexation records by keyword
}
